Large change: changed time stepping for 3D u and tracers
Both time steps in 3D u/v momentum and tracers arrays are set to
the current climatology entry matching that time step. This now
matches the behavior of ubar/vbar and zeta. This should make it so
that time steps match between online and offline for a perfect
offline simulation for 3D u/v and salt/temp tracers.