fix boundary issue and working on gls/tke/AKv
The changes to ini_fields appear to have fixed ubar/vbar and
salt/temp initial boundary condition problem where it was being
applied in the offline case but shouldn't have been.
Now trying to get the turbulence/mixing parameters to all match
the online case perfectly. Moved the updating of the climatology
to be before setting the mixing parameters in set_data.F to help
address this.