Integrating flags into log file and finishing salt/temp
All offline flags should now appear in the logfile if properly
activated. Also the ability to input salt/temp through
climatology for an offline tracer run has been put back in. This
does not impact an offline tracer run but they can be input to be
used for example with offline floats and weathering, etc.